Widen corruptSingleChar complement to uint256
~CMASK_STRING_LITERAL_TAIL complemented at uint128 width, so once
zero-extended the high 128 bits were clear and any rolled byte >= 0x80
kept the loop spinning until it landed in {0x00-0x1F, 0x7F}. Complement
at uint256 width instead so the corruption alphabet matches the
documented "not a valid character in a string literal and not a double
quote" across the full byte range, and bytes >= 0x80 already at the
index are recognised as invalid and kept.

+++ src/lib/mut/LibConformString.sol
@@ -105,7 +105,7 @@ library LibConformString {
         uint256 char = uint256(uint8(bytes(str)[index]));
         uint256 seed = 0;
         // forge-lint: disable-next-line(unsafe-typecast,incorrect-shift)
-        while (1 << char & ~CMASK_STRING_LITERAL_TAIL == 0 || char == uint8(bytes1("\""))) {
+        while (1 << char & ~uint256(CMASK_STRING_LITERAL_TAIL) == 0 || char == uint8(bytes1("\""))) {
             assembly ("memory-safe") {
                 mstore(0, char)
                 mstore(0x20, seed)
